pyproject.toml
[project]
name = "orga-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aerich[toml]>=0.9.0",
    "robyn>=0.68.0",
    "tortoise-orm[accel]>=0.25.1",
    "ulid>=1.1",
    "python-dotenv",
    "asyncpg",
    "orjson>=3.9.15",
]

[tool.aerich]
tortoise_orm = "src.database.TORTOISE_ORM"
location = "./src/database/migrations"
src_folder = "./."
